+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0] - 2020-10-25
+
+### New features
+- winter.web now supports problem annotation and handling exceptions accordingly with RFC7807  
+### Changed
+- Setting winter requires an explicit call to setup functions
+
 ## [4.1.1] - 2020-10-06
 
 ### Fixed